Move client snapshotter to separate package
